Move metrics module under main package
